requirements.txt
Django>=4.2
pymongo
python-dotenv
